Migrate from deprecated database_provides/requires to data_interfaces
