+++ src/relations/database_provides.py
@@ -6,7 +6,7 @@
 import json
 import logging
 
-from charms.data_platform_libs.v0.database_provides import (
+from charms.data_platform_libs.v0.data_interfaces import (
     DatabaseProvides,
     DatabaseRequestedEvent,
 )
