+++ src/relations/database_requires.py
@@ -7,7 +7,7 @@ import json
 import logging
 from typing import Dict, Optional
 
-from charms.data_platform_libs.v0.database_requires import (
+from charms.data_platform_libs.v0.data_interfaces import (
     DatabaseCreatedEvent,
     DatabaseEndpointsChangedEvent,
     DatabaseRequires,
